requirements.txt
fastapi
uvicorn[standard]
openai
langchain-community
langchain-openai
langgraph
annotated-types
requests
typing-inspection
typing-extensions